requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
chardet>=5.0.0
